Remove ExternalIdDao.
